Add support for categorical data in multivariate analysis
To specify a metadata row as a multivariate, replace the original row
header with:
factor(reference_level): metadata_row

The script will parse out the name and reference level, and treat that
row as categorical data. Note that the script can't make intelligent
guesses about which metadata rows are feature and which are numerical.

Rows without the "factor" annotation will be treated as numerical.